LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/path-objects/2d.path.roundrect.nonfinite.worker-expected.txt


PASS roundRect() with Infinity/NaN is ignored
